+++ crates/zng-task/src/channel/ipc.rs
@@ -1,11 +1,12 @@
 use std::io;
 
 use serde::{Deserialize, Serialize};
+use zng_time::Deadline;
 
 use crate::channel::ChannelError;
 
 /// The transmitting end of an IPC channel.
-/// 
+///
 /// Use [`ipc_channel`](self::ipc_channel) to declare a new channel.
 #[derive(Serialize, Deserialize)]
 pub struct IpcSender<T> {
@@ -13,39 +14,76 @@ pub struct IpcSender<T> {
 }
 impl<T: IpcValue> Clone for IpcSender<T> {
     fn clone(&self) -> Self {
-        Self { sender: self.sender.clone() }
+        Self {
+            sender: self.sender.clone(),
+        }
     }
 }
 impl<T: IpcValue> IpcSender<T> {
-    /// Send a value into the channel..
+    /// Send a value into the channel.
+    /// 
+    /// IPC channels are unbounded, this never blocks.
     pub fn send(&self, msg: T) -> Result<(), ChannelError> {
-        self.sender.send(msg).map_err(ChannelError::other) // !!: TODO should this be async? other zng_task channels are
+        self.sender.send(msg).map_err(ChannelError::other)
     }
 }
 
 /// The receiving end of an IPC channel.
+///
+/// Use [`ipc_channel`](self::ipc_channel) to declare a new channel.
 #[derive(Serialize, Deserialize)]
 pub struct IpcReceiver<T> {
-    recv: ipc_channel::ipc::IpcReceiver<T>,
+    recv: Option<ipc_channel::ipc::IpcReceiver<T>>,
 }
 impl<T: IpcValue> IpcReceiver<T> {
     /// Wait for an incoming value from the channel associated with this receiver.
-    pub fn recv(&self) -> Result<T, ChannelError> {
-        let r = self.recv.recv()?;
+    ///
+    /// Returns an error if all senders have been dropped.
+    pub async fn recv(&mut self) -> Result<T, ChannelError> {
+        let recv = self.recv.take().unwrap();
+        let (recv, r) = crate::wait(move || {
+            let r = recv.recv();
+            (recv, r)
+        }).await;
+        self.recv = Some(recv);
+        Ok(r?)
+    }
+    
+    /// Block for an incoming value from the channel associated with this receiver.
+    ///
+    /// Returns an error if all senders have been dropped or the `deadline` is reached.
+    pub async fn recv_deadline(&mut self, deadline: impl Into<Deadline>) -> Result<T, ChannelError> {
+        match crate::with_deadline(self.recv(), deadline).await {
+            Ok(r) => r,
+            Err(_) => Err(ChannelError::Timeout),
+        }
+    }
+
+    /// Block for an incoming value from the channel associated with this receiver.
+    pub fn recv_blocking(&self) -> Result<T, ChannelError> {
+        let r = self.recv.as_ref().unwrap().recv()?;
         Ok(r)
     }
 
-    // !!: TODO deadline API, async?
+    /// Block for an incoming value from the channel associated with this receiver.
+    ///
+    /// Returns an error if all senders have been dropped or the `deadline` is reached.
+    pub fn recv_deadline_blocking(&self, deadline: impl Into<Deadline>) -> Result<T, ChannelError> {
+        match deadline.into().time_left() {
+            Some(d) => Ok(self.recv.as_ref().unwrap().try_recv_timeout(d)?),
+            None => Err(ChannelError::Timeout),
+        }
+    }
 }
 
 /// Create an IPC channel.
-/// 
+///
 /// Channel is unbounded, that is, equivalent to [`channel::unbounded`], but capable of communication with another process
 /// by serializing and deserializing messages.
-/// 
+///
 /// Note that the channel endpoints can also be send over IPC, the first channel is setup by [`process::Worker`]. You
 /// can also use the [`ipc_channel`] crate to setup the first channel with a custom worker process.
-/// 
+///
 /// [`channel::unbounded`]: crate::channel::unbounded
 /// [`process::Worker`]: crate::process::Worker
 /// [`ipc_channel`]: https://docs.rs/ipc-channel/latest/ipc_channel/ipc/struct.IpcOneShotServer.html
@@ -54,7 +92,7 @@ where
     T: IpcValue,
 {
     let (s, r) = ipc_channel::ipc::channel()?;
-    Ok((IpcSender { sender: s }, IpcReceiver { recv: r }))
+    Ok((IpcSender { sender: s }, IpcReceiver { recv: Some(r) }))
 }
 
 /// Represents a type that can be an input and output of IPC channels.
@@ -69,7 +107,7 @@ where
 /// Types need to be `serde::Serialize + serde::de::Deserialize + Send + 'static` to auto-implement this trait,
 /// if you want to send an external type in that does not implement all the traits
 /// you may need to declare a *newtype* wrapper.
-#[diagnostic::on_unimplemented(note = "`IpcValue` is implemented for all `T: Debug + Serialize + Deserialize + Send + 'static`")]
+#[diagnostic::on_unimplemented(note = "`IpcValue` is implemented for all `T: Serialize + Deserialize + Send + 'static`")]
 pub trait IpcValue: serde::Serialize + for<'d> serde::de::Deserialize<'d> + Send + 'static {}
 
 impl<T: serde::Serialize + for<'d> serde::de::Deserialize<'d> + Send + 'static> IpcValue for T {}
@@ -78,6 +116,15 @@ impl From<ipc_channel::ipc::IpcError> for ChannelError {
     fn from(value: ipc_channel::ipc::IpcError) -> Self {
         match value {
             ipc_channel::ipc::IpcError::Disconnected => ChannelError::Disconnected,
+            e => ChannelError::other(e),
+        }
+    }
+}
+impl From<ipc_channel::ipc::TryRecvError> for ChannelError {
+    fn from(value: ipc_channel::ipc::TryRecvError) -> Self {
+        match value {
+            ipc_channel::ipc::TryRecvError::IpcError(ipc_channel::ipc::IpcError::Disconnected) => ChannelError::Disconnected,
+            ipc_channel::ipc::TryRecvError::Empty => ChannelError::Timeout,
             e => ChannelError::other(e),
         }
     }